refactor: share Recharts tooltip helpers
