+++ frontend/src/dashboard/widgets/spending-comparison/SpendingComparisonChart.tsx
@@ -17,7 +17,7 @@ import {
   getRechartsTooltipPoint,
   getRechartsTooltipPointer,
   type RechartsTooltipState,
-} from '@/dashboard/utils/rechartsTooltip'
+} from '@/components/charts/rechartsTooltip'
 
 import { SpendingComparisonTooltipContent } from './SpendingComparisonTooltipContent'
