+++ frontend/tests/components/charts/rechartsTooltip.test.ts
@@ -10,7 +10,7 @@ import {
   getRechartsTooltipPoint,
   getRechartsTooltipPointer,
   type RechartsTooltipState,
-} from '@/dashboard/utils/rechartsTooltip'
+} from '@/components/charts/rechartsTooltip'
 
 type TestPoint = {
   label: string
